Relocate F680 unit tests
